:recycle: Refactor project structure by main parts of generation
