--- internal/ime_survey/app_begin_generation.go
@@ -1,15 +1,0 @@
-package ime_survey
-
-import (
-	"gopkg.in/AlecAivazis/survey.v1"
-)
-
-func AskBeginGeneration(launchGeneration *bool) error {
-	answer := false
-	err := survey.AskOne(&survey.Confirm{Message: "Begin generation?"}, &answer, nil)
-	if err != nil {
-		return err
-	}
-	*launchGeneration = answer
-	return nil
-}
